Add files via upload
Hypothesis Testing Exercise
A F&B manager wants to determine whether there is any
significant difference in the diameter of the cutlet between two
units. A randomly selected sample of cutlets was collected from
both units and measured? Analyze the data and draw inferences
at 5% significance level. Please state the assumptions and tests
that you carried out to check validity of the assumptions.